Cast day_of_week to int in get_class_timetable RPC
timetable_slots.day_of_week is actually stored as TEXT (e.g. "1"),
not INT — caused "structure of query does not match function result
type" since the declared RETURNS TABLE column is INT.

+++ scripts/timetable-public-rpc.sql
@@ -34,7 +34,7 @@ AS $$
 BEGIN
   RETURN QUERY
   SELECT
-    ts.day_of_week,
+    ts.day_of_week::int,
     ts.period_number,
     ts.slot_type,
     ts.subject_name,
